refactor(observability): split runtime concerns
