+++ packages/observability-plugin/src/chrome-devtool.ts
@@ -1,7 +1,10 @@
-import type { ObservabilityRuntimePlugin } from './core';
-import { createObservability, type ObservabilityPluginOptions } from './core';
+import { createObservability } from './core';
+import type {
+  ObservabilityPluginOptions,
+  ObservabilityRuntimePlugin,
+} from './type';
 
-export type { ObservabilityPluginOptions } from './core';
+export type { ObservabilityPluginOptions } from './type';
 
 export function ChromeObservabilityPlugin(
   options: ObservabilityPluginOptions = {},
